src/main/java/edu/escuelaing/arsw/ASE/util/LineCounter.java
package edu.escuelaing.arsw.ASE.util;

import java.io.BufferedReader;
import java.io.File;
import java.io.FileReader;
import java.io.IOException;

public class LineCounter {

    public static int countPhysicalLines(File file) throws IOException {
        int lines = 0;
        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
            while (reader.readLine() != null) {
                lines++;
            }
        }
        return lines;
    }

    public static int countLOC(File file) throws IOException {
        int loc = 0;
        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
            String line;
            while ((line = reader.readLine()) != null) {
                line = line.trim();
                if (!line.isEmpty() && !line.startsWith("//") && !line.startsWith("/*")) {
                    loc++;
                }
            }
        }
        return loc;
    }
}
